Move dashboard directory into new home directory
